requirements.txt
spacy>=3.8.0
sentence-transformers>=2.2.2
pandas>=2.1.3
numpy>=1.26.2
pyodbc>=5.0.1
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
scikit-learn>=1.3.2
pyarrow>=14.0.0